split wrapped module into external and internal
