+++ rust/app/src/wrappers/external_references/src/polywrap/external/global_functions/test_external_global_function.rs
@@ -6,7 +6,7 @@ use serde_json::json;
 use crate::polywrap::external::module::wrap_module::get_external_module_or_panic;
 use crate::polywrap::wrap_manifest::WrapManifest;
 use crate::polywrap::resources::ExternalResource;
-use crate::polywrap::wrapped::StringWrapped;
+use crate::polywrap::internal::wrapped::StringWrapped;
 
 pub async fn testExternalGlobalFunction(
   arg: String,
