+++ rust/app/src/wrappers/external_references/src/polywrap/external/mod.rs
@@ -1,5 +1,6 @@
 pub mod global_functions;
 pub mod classes;
 pub mod module;
+pub mod wrapped;
 pub use classes::*;
 pub use global_functions::*;
