+++ rust/app/src/wrappers/external_references/src/polywrap/external/wrapped/mod.rs
@@ -1,0 +1,2 @@
+pub mod test_external_class;
+pub use test_external_class::test_external_class_wrapped::TestExternalClassWrapped;
