+++ rust/app/src/wrappers/external_references/src/polywrap/external/wrapped/test_external_class/invoke.rs
