+++ rust/app/src/wrappers/external_references/src/polywrap/internal/invoke_global_function.rs
@@ -15,7 +15,8 @@ use crate::{
   testInvokeExternalInstanceMethod,
        
 };
-use crate::polywrap::wrapped::{TestExternalClassWrapped, StringWrapped};
+use crate::polywrap::external::wrapped::TestExternalClassWrapped;
+use crate::polywrap::internal::wrapped::StringWrapped;
 use crate::polywrap::external::TestExternalClass;
