+++ rust/app/src/wrappers/external_references/src/polywrap/internal/mod.rs
@@ -1,4 +1,5 @@
 pub mod invoke_class_method;
 pub mod invoke_global_function;
+pub mod wrapped;
 pub use invoke_class_method::invoke_class_method;
 pub use invoke_global_function::invoke as invoke_global_function;
