+++ rust/app/src/wrappers/external_references/src/polywrap/internal/wrapped/mod.rs
@@ -1,6 +1,4 @@
 pub mod string_wrapped;
-pub mod test_external_class;
 pub mod test_object_getter;
 pub use string_wrapped::StringWrapped;
-pub use test_external_class::test_external_class_wrapped::TestExternalClassWrapped;
 pub use test_object_getter::test_object_getter_wrapped::TestObjectGetterWrapped;
