+++ rust/app/src/wrappers/external_references/src/polywrap/mod.rs
@@ -1,7 +1,5 @@
 pub mod external;
 pub mod internal;
 pub mod resources;
-pub mod wrapped;
 pub mod wrap_manifest;
 pub mod exports;
-pub use exports::*;
